+++ src/sound/YMF262.hh
@@ -155,7 +155,7 @@ private:
 		const EmuTime& time, const EmuDuration& sampDur);
 
 	// Resample
-	virtual bool generateInput(float* buffer, unsigned num);
+	virtual bool generateInput(int* buffer, unsigned num);
 
 	void callback(byte flag);
